fix gotcha when using symlink for script
Thanks for this handy tool. I've got a pile of textils2markdown conversion to work though and this is making it much easier.

This pull request is for a minor tweak needed to do to allow the tool to find it's libraries when textile2markdown is a symlink.

My setup was: clone in my project's dir and add a symlink to tool in my $HOME/bin dir.